+++ back-end/src/routes/api/books.js
@@ -1,23 +1,6 @@
 const router = require('express').Router();
-const request = require('request');
-const createController = require('../createController');
+const controller = require('../../controllers/api/books');
 
-router.post(
-  '/getBooks',
-  createController(async (req, res) => {
-    const query = req.body.bookQuery.split(/\s*\s/).join('+');
-
-    request(
-      'https://www.googleapis.com/books/v1/volumes?q=' + query,
-      { json: true },
-      (err, response, body) => {
-        if (err) {
-          return console.log(err);
-        }
-        return res.json(response);
-      },
-    );
-  }),
-);
+router.post('/getBooks', controller.get);
 
 module.exports = router;
